+++ projects/rxjs-idsets/CHANGELOG.md
@@ -3,5 +3,6 @@
 ## 1.0.0
 - Initial public version
 - TODO:
+  - Add a version of the CategorizedIdSet that publishes readonly category IDSets to make category cleanup possible.
   - Improve documentation
   - Add more examples